client/requirements.txt
numpy
pypng
wxPython
PyPubSub
matplotlib
